requirements.txt
flask
flask-sqlalchemy
flask-migrate
scapy
python-nmap
requests
pandas
